SymbolTable: move to analyses sub package
